Added `CARGO_MAKE_TEST_BINARY_FILTER` filter variable for coverage
When coverage is collected, only the binaries that match the regex
specified by this variable are run. By default, crate unit test binaries
are all prefixed with the crate name. Integration tests are given names
based on their file name without the crate name prefix, so users have to
override this variable for those binaries to be run during coverage.

Issue #59