Ensure that the specified rustup toolchain exists before using it
Given a config like this

    [tasks.test]
    command = "cargo"
    args = [ "build" ]
    toolchain = "alarm"

you get a fairly unintuitive error output like this:

```
[cargo-make] INFO - Running Task: test
[cargo-make] ERROR - Error while executing command, exit code: 1
[cargo-make] WARN - Build Failed.
```

This can be quite confusing especially for people who are new to a
project and might've missed installing the specified rustup toolchain
before building the project with `cargo-make`.

This patch ensures in the `toolchain`-subcrate that the toolchain
actually exists (by checking the exit-code of `rustup run <toolchain>
true`) before running the actual task. If the toolchain doesn't exist,
an error like this will be thrown:

```
[cargo-make] INFO - Running Task: test
[cargo-make] ERROR - Missing toolchain alarm! Please install it using rustup.
[cargo-make] WARN - Build Failed.
```