Allow for explicit workspace flow in sub flows
Commit fa929b116f78 ("Fix workspace detection for sub flows") caused a
regression when having a root workspace with tasks that relies on each
member to define CARGO_MAKE_CARGO_BUILD_TEST_FLAGS with unique set of
features.

Determine if workspace flow is explicitly set in the task when planning
the execution and take this setting into account when running in a sub
flow.