Provide more detailed error message when env-script fails
There is limited information given in the error output when an
env-script return a non-zero exit code, which can be quite confusing as
there is no context of where it failed (unless verbose output is turned
on).

Avoid confusion by providing more context about the environment when the
script fails.

BEFORE:
 > cargo make
 [cargo-make] INFO - cargo make 0.32.5
 [cargo-make] ERROR - Error while executing command, exit code: 128
 [cargo-make] WARN - Build Failed.

AFTER:
 > cargo make
 [cargo-make] INFO - cargo make 0.32.5
 [cargo-make] ERROR - Error while evaluating script for env: GIT_ROOT, exit code: 128
 script:
 [
     "git rev-parse --show-toplevel",
 ]
 stdout:

 stderr:
 fatal: not a git repository (or any of the parent directories): .git

 [cargo-make] WARN - Build Failed.